Import CRYPTOGAMS ASM code (from OpenSSL 3.0.0-beta1)
The CRYPTOGAMS code is suitably licensed for use in OpenConnect under
LGPLv2.1, and gives us a 40% speedup to ESP AES-SHA1 encryption.

However, not everything is in the standalone CRYPTOGAMS repository, so
we have to import from OpenSSL itself for now, which means the licence
is incompatible.

Once https://github.com/dot-asm/cryptogams/issues/7 is resolved, we can
do this for real. But for now it's worth having it to experiment with.

Really needs SHA256 too...

Signed-off-by: David Woodhouse <dwmw2@infradead.org>