GnuTLS challenge signing
This is the GnuTLS implementation for signing the challenge. Variables
added to store the paths for the second set of credentials.

Challenge response comprises two elements: an identity and signature.
The identity is the user certificate, represented as a PEM-encoded PKCS7
certificate chain. The challenge is signed with the user certificate
and submitted back to the server as a PEM-encoded signature.

Signed-off-by: Tom Carroll <incentivedesign@gmail.com>