Factored out key+cert match mechanism
Previously there were two alternates to check that the key matched the
certificate. Alternative one was employed if the key was a x509_privkey;
alternative two was used if the key was an abstract privkey. By
preferring the abstract privkey, the code can be deduplicated and
simplified. The adoption was modeled after GnuTLS 3.6.13
_gnutls_check_key_cert_match, which GnuTLS uses to verify that key and
cert are a pair. Match was factored out of load_certificate and put into
a separate function.

Signed-off-by: Tom Carroll <incentivedesign@gmail.com>