attempt to coax server to accept a larger MRU by nak-offering our MTU if it's greater than the server's MRU
We only attempt this once before giving up.

Tested on F5, and this works: we nak-offer a larger MRU, and the server
responds with a new CONFACK requesting the same value.

The F5 server often offers erroneously low MRU values, too small for IPv6
(minimum MTU of 1280), and is consistently strict about not accepting
incoming IP packets which are even 1 byte larger than its MRU, so this is
useful. 🍺

Signed-off-by: Daniel Lenski <dlenski@gmail.com>