rename CLI options containing 'dtls' to 'udp', and 'csd' to 'trojan' where applicable
Many of the CLI options with 'dtls' in the string are in fact applicable to
all protocols' UDP transports.  Likewise for 'csd' options, which are
applicable to all protocols' Trojan binary/script execution.

This patch renames all of the options, but retains the old names as
redundant/duplicate options for backwards-compatibility.

Signed-off-by: Daniel Lenski <dlenski@gmail.com>