Add DTLS interaction to ppp_tcp_mainloop()
Based on the logic in gpst_mainloop() for ESP deferral, it waits and does
nothing while DTLS is still active (including trying to connect).

Only if/when DTLS is out of the picture does it start trying to establish
PPP for itself. If the PPP state is still PPPS_DEAD, that means that the
protocol's ->tcp_connect() method never actually established the tunnel,
so call it directly; falling back to a full ssl_reconnect() if needed.

The protocol's ->tcp_connect() method is then expected to establish the
tunnel *and* call ppp_start_tcp_mainloop() which will do one pass
through the state machine and shift from PPPS_DEAD to PPPS_ESTABLISH.

Subsequent calls to ppp_tcp_mainloop() see a state other than PPPS_DEAD
(as well as DTLS disabled) and call directly through to the core PPP
mainloop for processing.

Signed-off-by: David Woodhouse <dwmw2@infradead.org>