Fix timeout handling for DTLS handshake retries
We weren't actually waking the mainloop up when the TLS library wanted
to (re)send a DTLS handshake. Plumb the timeout pointer through
dtls_try_handshake() to fix that.

Also feed it to dtls_reconnect(), which helps us ensure that we wake up
after dtls_attempt_period() to try again.

Export dtls_reconnect() while we're at it, since PPP will want to be
able to invoke it.

Signed-off-by: David Woodhouse <dwmw2@infradead.org>