Fortinet: don't keep retrying if cookie is invalid on reconnect
The fortinet_configure() requests return 302 redirects to '/remote/login' if
the auth session/cookie is no longer valid. We should detect this and
return -EPERM rather than -EINVAL, so that ssl_reconnect() doesn't keep
trying to reconnect.

NB: Detecting this redirect is perhaps a bit harder than it should be,
because do_https_request() returns 0, rather than the real HTTP status code
(e.g.  302), in the case of an successful-but-unfetched redirect.

Signed-off-by: Daniel Lenski <dlenski@gmail.com>