Exclude cookies from F5's ppp_tls_connect_req
This buffer is used to initiate the connection over either TLS or DTLS.
Cookies are not needed for it to succeed, and can potentially grow without bound,
which would make it too big to fit in a single DTLS packet (ick, HTTP over DTLS).
Therefore we exclude them.

Don't blame me. I didn't design this.

Signed-off-by: Daniel Lenski <dlenski@gmail.com>