Fix f5-auth-and-config tests not to depend on cookies
Tests depended on cookies propagating all the way to the tunnel
connection ('GET /tmyvpn'). Real servers don't need to see the cookie
at this point though, and for the DTLS connection we don't *want* to
include it because the resulting packet could end up too large. So
don't require it.

Signed-off-by: Daniel Lenski <dlenski@gmail.com>