Fix PPP reconnect
In rebasing and cleaning up, I appear to have screwed up the PPP reconnect
state machine. The UDP mainloop runs first, starts a connection, which
is immediately torn down by the TCP mainloop because of the stale PPP
state from the previous connection.

Fix it by resetting the PPP state when attempting the DTLS reconnection
after a pause (which we determine by observing that the state isn't
PPPS_DEAD but the ssl_fd isn't actually connected).

The Fortinet UDP mainloop wants the same trick. And I want to consolidate
the two anyway.

[DL copied to Fortinet UDP mainloop.]

Signed-off-by: David Woodhouse <dwmw2@infradead.org>
Signed-off-by: Daniel Lenski <dlenski@gmail.com>