Partial fix for Fortinet auth
We have to fill in form->auth_id in all cases, and were forgetting in
the synthesised password form.

Also, on a redirect we would get an empty resp_buf from do_http_request()
which would cause a crash when we dereference it. Don't do that.

We still need some more work here (and a webview at least for the GUI
case) but this at least fixes the worst of it.

Signed-off-by: David Woodhouse <dwmw2@infradead.org>