Don't try to set MTU on tunnel interface within (lib)openconnect itself
This has been the responsibility of the vpnc-script for a very long time,
but the MTU-setting code was retained in openconnect for compatibility with
ancient versions of the vpnc-script.  The MTU-setting code causes noisy
errors when running OpenConnect as non-root, and is simply redundant with
all vpnc-scripts since at least 2009 (see
https://gitlab.com/openconnect/vpnc-scripts/commit/02b674554db6cf662283ed74941564ccb9c5a8a4).

We will need to refine the MTU-setting code in the future (see
https://gitlab.com/openconnect/openconnect/-/issues/244#note_576191228), but
for now we should remove this code.

Signed-off-by: Daniel Lenski <dlenski@gmail.com>