Handle default Legacy IP route for oNCP as we do for GP
If a default Legacy IP route is specified as a "split"-include route, then
we should use that route as the Legacy IP netmask, and add the original
netmask as a split (if it's anything other than 255.255.255.255 or /32).

See https://gitlab.com/openconnect/openconnect/-/merge_requests/118 for the
rationale.

This patch factors out the function finalize_netmask_fixing_default_route_as_split()
from gpst.c and makes it a global internal function, so that the mechanism
can be shared.

Signed-off-by: Daniel Lenski <dlenski@gmail.com>