Make xmlnode_get_val() trim whitespace from XML node content
This is motivated by the fact that GlobalProtect portals often put extra
whitespace in the software version tag ('<version>   5.1.2-3
</version>'), and there is no case where we actually want to keep
leading/trailing whitespace.

The function fetch_and_trim() from xml.c already handles the needed
whitespace-trimming behavior, but we need to *duplicate* it (as
xmlnode_get_trimmed_val) in order to use it in both openconnect (the
application) and libopenconnect (the library), without making it a public
function of the library.

Signed-off-by: Daniel Lenski <dlenski@gmail.com>