nx: Remove TODO and debug prints regarding headers
So far, we have only one encountered one NX server that was enforcing a
certain client version and thus User-Agent header. Remove the TODO
pondering whether to add it or not and some leftover debug-print.

If we realise later, that this is more common, we can always add such a
header. In initial tests, the server did not do a 1:1 string comparison,
so we could still be transparent about being OpenConnect by prepending
or appending OpenConnect to the "approved" User-Agent string.

Signed-off-by: Andreas Gnau <rondom@rondom.de>